Build delete update

+++ src/main/java/com/example/SearchEngine/invertedIndex/service/ranking/Bm25Ranker.java
@@ -11,7 +11,7 @@ public class Bm25Ranker implements Ranker {
     private final Double K = 2.0;
     private final Double B = 0.75;
 
-    private Double Bm25Ranker(Double df, Double tf, Integer documentId, String schemaName) {
+    private Double Bm25Ranker(Double df, Double tf, Long documentId, String schemaName) {
         Double N = Double.valueOf(CollectionInfo.getNumberOfDocument(schemaName));
         Double documentLength = CollectionInfo.getDocumentLength(schemaName, documentId);
         Double avgdl = CollectionInfo.getDocumentsTotalLength(schemaName) / N;
@@ -22,9 +22,9 @@ public class Bm25Ranker implements Ranker {
 
 
     @Override
-    public HashMap<Integer, Double> calculateScore(String schemaName, HashMap<Integer, HashMap<String, Double>> documents) {
-        HashMap<Integer, Double> currentDocumentsScores = new HashMap();
-        for (int documentId : documents.keySet()) {
+    public HashMap<Long, Double> calculateScore(String schemaName, HashMap<Long, HashMap<String, Double>> documents) {
+        HashMap<Long, Double> currentDocumentsScores = new HashMap();
+        for (Long documentId : documents.keySet()) {
             Double df = (double) documents.size();
             Double tf = documents.get(documentId).get("total");
             currentDocumentsScores.put(documentId, Bm25Ranker(df, tf, documentId, schemaName));
@@ -33,12 +33,12 @@ public class Bm25Ranker implements Ranker {
     }
 
     @Override
-    public List<Integer> rankDocuments(HashMap<Integer, Double> documentsScores) {
+    public List<Long> rankDocuments(HashMap<Long, Double> documentsScores) {
 
-        List<Map.Entry<Integer, Double>> documents = new ArrayList<>(documentsScores.entrySet());
+        List<Map.Entry<Long, Double>> documents = new ArrayList<>(documentsScores.entrySet());
         Collections.sort(documents, (a, b) -> b.getValue().compareTo(a.getValue()));
-        List<Integer> sortedKeys = new ArrayList<>();
-        for (Map.Entry<Integer, Double> entry : documents) {
+        List<Long> sortedKeys = new ArrayList<>();
+        for (Map.Entry<Long, Double> entry : documents) {
             sortedKeys.add(entry.getKey());
         }
         return sortedKeys;
